+++ src/fhir-models/Bundle.ts
@@ -133,8 +133,8 @@ export class Bundle extends Resource implements IBase {
    * @param optSourceField - Optional data source field (e.g. `<complexTypeName>.<complexTypeFieldName>`); defaults to Bundle
    * @returns Bundle data model or undefined for `Bundle`
    */
-  public static override parse(sourceJson: JSON.Object, optSourceField?: string): Bundle | undefined {
-    if (!isDefined<JSON.Object>(sourceJson) || (JSON.isJsonObject(sourceJson) && isEmpty(sourceJson))) {
+  public static override parse(sourceJson: JSON.Value, optSourceField?: string): Bundle | undefined {
+    if (!isDefined<JSON.Value>(sourceJson) || (JSON.isJsonObject(sourceJson) && isEmpty(sourceJson))) {
       return undefined;
     }
     const source = isDefined<string>(optSourceField) ? optSourceField : 'Bundle';