+++ src/fhir-models/PractitionerRole.ts
@@ -116,8 +116,8 @@ export class PractitionerRole extends DomainResource implements IBase {
    * @param optSourceField - Optional data source field (e.g. `<complexTypeName>.<complexTypeFieldName>`); defaults to PractitionerRole
    * @returns PractitionerRole data model or undefined for `PractitionerRole`
    */
-  public static override parse(sourceJson: JSON.Object, optSourceField?: string): PractitionerRole | undefined {
-    if (!isDefined<JSON.Object>(sourceJson) || (JSON.isJsonObject(sourceJson) && isEmpty(sourceJson))) {
+  public static override parse(sourceJson: JSON.Value, optSourceField?: string): PractitionerRole | undefined {
+    if (!isDefined<JSON.Value>(sourceJson) || (JSON.isJsonObject(sourceJson) && isEmpty(sourceJson))) {
       return undefined;
     }
     const source = isDefined<string>(optSourceField) ? optSourceField : 'PractitionerRole';